backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
openai
google-generativeai
pyyaml
python-multipart
httpx
anthropic
